Fix bugs

+++ backend/app/app/models/phrase_fullsearch.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Date, DECIMAL
+from sqlalchemy import Column, Integer, String, Date, DECIMAL, BigInteger
 
 from app.db.base_class import Base
 
@@ -6,7 +6,7 @@ from app.db.base_class import Base
 class PhraseFullSearch(Base):
     id = Column(Integer, primary_key=True, index=True)
     keyword = Column(String, index=True, nullable=False)
-    number_results = Column(Integer, nullable=False, default=0)
+    number_results = Column(BigInteger, nullable=False, default=0)
     cpc = Column(DECIMAL, nullable=False, default=0.0)
     competition = Column(Integer, nullable=False, default=0)
     search_volume = Column(Integer, nullable=False, default=0)
